pyproject.toml
[tool.poetry]
name = "test-school-it"
version = "0.1.0"
description = ""
authors = ["Oleg <145393954+Ole6G@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg = "^3.2.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
